cloud-functions/requirements.txt
functions-framework==3.*
google-cloud-compute==1.26.0
google-auth
